% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{mlr3_nuisance_fit}
\alias{mlr3_nuisance_fit}
\title{Build an mlr3 nuisance callback for fit_gmm}
\usage{
mlr3_nuisance_fit(specs)
}
\arguments{
\item{specs}{A named list of nuisance specifications.}
}
\value{
A callback accepted by [fit_gmm()].
}
\description{
Build an mlr3 nuisance callback for fit_gmm
}
