% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{nuisance_spec}
\alias{nuisance_spec}
\title{Specify one mlr3 nuisance learner}
\usage{
nuisance_spec(
  learner,
  target,
  features,
  groups = NULL,
  predict_type = "response"
)
}
\arguments{
\item{learner}{An `mlr3` learner or `mlr3tuning` auto-tuner.}

\item{target}{One target column name or a selector function.}

\item{features}{Feature column names or a selector function.}

\item{groups}{Optional one or more clustering column names used for tuning.}

\item{predict_type}{The mlr3 prediction type, normally `"response"`.}
}
\value{
A `fullsampleDML_nuisance_spec` object.
}
\description{
Specify one mlr3 nuisance learner
}
