% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullsampleDML-package.R
\docType{package}
\name{fullsampleDML-package}
\alias{fullsampleDML}
\alias{fullsampleDML-package}
\title{fullsampleDML: full-sample debiased GMM}
\description{
Native R implementation of full-sample nuisance estimation, generic
user-defined GMM scores, and arbitrary-way clustered inference.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kaichengchen/fullsampleDML-r}
  \item \url{https://arxiv.org/abs/2602.11333}
}

}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Kaicheng Chen \email{chenkaicheng@sufe.edu.cn}
  \item Harold D. Chiang \email{hdchiang@wisc.edu}
}

}
\keyword{internal}
